+++ iOS/Records/HealthKit/Models/HealthRecord.swift
@@ -17,34 +17,3 @@ extension HealthRecord {
         "\(sample.sourceRevision.source.name) \(sample.sourceRevision.source.bundleIdentifier)"
     }
 }
-
-struct Carbs: @MainActor HealthRecord {
-    let id = UUID()
-    let value: Double
-    let sample: HKQuantitySample
-}
-
-struct Glucose: @MainActor HealthRecord {
-    let id = UUID()
-    let value: Double
-    let sample: HKQuantitySample
-}
-
-struct Insulin: @MainActor HealthRecord {
-    let id = UUID()
-    let value: Double
-    let type: InsulinType
-    let sample: HKQuantitySample
-    
-    var formattedValue: String {
-        value.truncatingRemainder(dividingBy: 1) == 0 ?
-        String(format: "%.0f", value) :
-        String(value)
-    }
-}
-
-struct Weight: @MainActor HealthRecord {
-    let id = UUID()
-    let value: Double
-    let sample: HKQuantitySample
-}
