+++ iOS/Records/HealthKit/Models/InsulinType.swift
@@ -6,21 +6,18 @@ enum InsulinType: String, Identifiable, Codable, CaseIterable {
     var id: String {
         rawValue
     }
-
+    
     var title: String {
         switch self {
         case .bolus: String(localized: "Bolus")
         case .basal: String(localized: "Basal")
         }
     }
-
+    
     var healthKitValue: Int {
         switch self {
-        case .bolus:
-            HKInsulinDeliveryReason.bolus.rawValue
-
-        case .basal:
-            HKInsulinDeliveryReason.basal.rawValue
+        case .bolus: HKInsulinDeliveryReason.bolus.rawValue
+        case .basal: HKInsulinDeliveryReason.basal.rawValue
         }
     }
 }
